Move `help` section to separate file
