+++ src/modules/metrics/constants/metric-helps.ts
@@ -1,0 +1,7 @@
+export const REDIS_METRIC_HELP = {
+  REQUESTS_TOTAL: 'Total Redis cache requests',
+  HITS_TOTAL: 'Total Redis cache hits',
+  MISSES_TOTAL: 'Total Redis cache misses',
+  ERRORS_TOTAL: 'Total Redis cache errors',
+  RESPONSE_TIME_SECONDS: 'Redis response time in seconds',
+};
